requirements.txt
# Azure SDKs
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-core==1.29.5

# OpenAI for embeddings
openai==1.12.0

# Data Source Connectors
atlassian-python-api==3.41.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2

# Utilities
python-dotenv==1.0.1
requests==2.31.0
tenacity==8.2.3
tiktoken==0.6.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Web Frameworks
flask==3.0.2
flask-cors==4.0.0
fastapi==0.109.2
uvicorn[standard]==0.27.1

# CLI
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Logging
structlog==24.1.0

# Testing (dev)
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality (dev)
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
